refactor(send): move selected-pool balance helper out of live-common
