+++ .changeset/send-balance-type.md
@@ -1,0 +1,7 @@
+---
+"@ledgerhq/live-send": minor
+"ledger-live-desktop": patch
+"live-mobile": patch
+---
+
+Move selected send-pool balance helper out of live-common into @ledgerhq/live-send
